# Improved performance for code2vec