--- a/dataset/AddContactDialog_actionPerformed_24.0_9.0_18.0_21.java
+++ b/dataset/AddContactDialog_actionPerformed_24.0_9.0_18.0_21.java
@@ -1,0 +1,22 @@
+public void actionPerformed(ActionEvent e)
+            {
+                if (groupCombo.getSelectedItem() != null
+                    && groupCombo.getSelectedItem().equals(newGroupString))
+                {
+                    CreateGroupDialog dialog
+                        = new CreateGroupDialog(parentDialog, false);
+                    dialog.setModal(true);
+                    dialog.setVisible(true);
+
+                    MetaContactGroup newGroup = dialog.getNewMetaGroup();
+
+                    if (newGroup != null)
+                    {
+                        groupCombo.insertItemAt(newGroup,
+                                groupCombo.getItemCount() - 2);
+                        groupCombo.setSelectedItem(newGroup);
+                    }
+                    else
+                        groupCombo.setSelectedIndex(0);
+                }
+            }