--- a/dataset/K6ClientCodegen_getAccept_21.0_7.0_15.0_26.java
+++ b/dataset/K6ClientCodegen_getAccept_21.0_7.0_15.0_26.java
@@ -1,0 +1,27 @@
+private static String getAccept(OpenAPI openAPI, Operation operation) {
+        String accepts = null;
+        String defaultContentType = "application/json";
+        Set<String> producesInfo = getProducesInfo(openAPI, operation);
+        if (producesInfo != null && !producesInfo.isEmpty()) {
+            ArrayList<String> produces = new ArrayList<>(producesInfo);
+            StringBuilder sb = new StringBuilder();
+            for (String produce : produces) {
+                if (defaultContentType.equalsIgnoreCase(produce)) {
+                    accepts = defaultContentType;
+                    break;
+                } else {
+                    if (sb.length() > 0) {
+                        sb.append(",");
+                    }
+                    sb.append(produce);
+                }
+            }
+            if (accepts == null) {
+                accepts = sb.toString();
+            }
+        } else {
+            accepts = defaultContentType;
+        }
+
+        return accepts;
+    }