--- a/extractor.py
+++ b/extractor.py
@@ -1,5 +1,5 @@
 import subprocess
-
+import tempfile
 
 class Extractor:
     def __init__(self, config, jar_path, max_path_length, max_path_width):
@@ -8,34 +8,46 @@ class Extractor:
         self.max_path_width = max_path_width
         self.jar_path = jar_path
 
-    def extract_paths(self, path):
-        command = ['java', '-cp', self.jar_path, 'JavaExtractor.App', '--max_path_length',
-                   str(self.max_path_length), '--max_path_width', str(self.max_path_width), '--file', path, '--no_hash']
-        process = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-        out, err = process.communicate()
-        output = out.decode().splitlines()
-        if len(output) == 0:
-            err = err.decode()
-            raise ValueError(err)
-        hash_to_string_dict = {}
-        result = []
-        for i, line in enumerate(output):
-            parts = line.rstrip().split(' ')
-            method_name = parts[0]
-            current_result_line_parts = [method_name]
-            contexts = parts[1:]
-            for context in contexts[:self.config.MAX_CONTEXTS]:
-                context_parts = context.split(',')
-                context_word1 = context_parts[0]
-                context_path = context_parts[1]
-                context_word2 = context_parts[2]
-                hashed_path = str(self.java_string_hashcode(context_path))
-                hash_to_string_dict[hashed_path] = context_path
-                current_result_line_parts += ['%s,%s,%s' % (context_word1, hashed_path, context_word2)]
-            space_padding = ' ' * (self.config.MAX_CONTEXTS - len(contexts))
-            result_line = ' '.join(current_result_line_parts) + space_padding
-            result.append(result_line)
-        return result, hash_to_string_dict
+    def extract_paths(self, code_string):
+        tmp = tempfile.NamedTemporaryFile(delete=False)
+        out = None
+        try:
+            tmp.write(bytes(code_string, encoding='utf-8'))
+            tmp.close()
+            #with open(tmp.name) as f:
+                #code = f.read()
+                #print(f'you passed the file\n {code}')
+                
+            command = ['java', '-cp', self.jar_path, 'JavaExtractor.App', '--max_path_length',
+                       str(self.max_path_length), '--max_path_width', str(self.max_path_width), '--file', tmp.name, '--no_hash']
+            print('Running {''.join(command)}')
+            process = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            out, err = process.communicate()
+            output = out.decode().splitlines()
+            if len(output) == 0:
+                err = err.decode()
+                raise ValueError(err)
+            hash_to_string_dict = {}
+            result = []
+            for i, line in enumerate(output):
+                parts = line.rstrip().split(' ')
+                method_name = parts[0]
+                current_result_line_parts = [method_name]
+                contexts = parts[1:]
+                for context in contexts[:self.config.MAX_CONTEXTS]:
+                    context_parts = context.split(',')
+                    context_word1 = context_parts[0]
+                    context_path = context_parts[1]
+                    context_word2 = context_parts[2]
+                    hashed_path = str(self.java_string_hashcode(context_path))
+                    hash_to_string_dict[hashed_path] = context_path
+                    current_result_line_parts += ['%s,%s,%s' % (context_word1, hashed_path, context_word2)]
+                space_padding = ' ' * (self.config.MAX_CONTEXTS - len(contexts))
+                result_line = ' '.join(current_result_line_parts) + space_padding
+                result.append(result_line)
+        finally:
+            return result, hash_to_string_dict
+
 
     @staticmethod
     def java_string_hashcode(s):