--- a/interactive_predict.py
+++ b/interactive_predict.py
@@ -2,6 +2,11 @@ import traceback
 
 from common import common
 from extractor import Extractor
+from pathlib import Path
+import inspect
+import time
+from numpy import mean, median, quantile
+import json
 
 SHOW_TOP_CONTEXTS = 10
 MAX_PATH_LENGTH = 8
@@ -16,6 +21,8 @@ class InteractivePredictor:
         model.predict([])
         self.model = model
         self.config = config
+        self.model.prepare()
+        self.results = {}
         self.path_extractor = Extractor(config,
                                         jar_path=JAR_PATH,
                                         max_path_length=MAX_PATH_LENGTH,
@@ -25,33 +32,39 @@ class InteractivePredictor:
         with open(input_filename, 'r') as file:
             return file.readlines()
 
-    def predict(self):
-        input_filename = 'Input.java'
-        print('Starting interactive prediction...')
-        while True:
-            print(
-                'Modify the file: "%s" and press any key when ready, or "q" / "quit" / "exit" to exit' % input_filename)
-            user_input = input()
-            if user_input.lower() in self.exit_keywords:
-                print('Exiting...')
-                return
-            try:
-                predict_lines, hash_to_string_dict = self.path_extractor.extract_paths(input_filename)
-            except ValueError as e:
-                print(e)
-                continue
-            raw_prediction_results = self.model.predict(predict_lines)
-            method_prediction_results = common.parse_prediction_results(
-                raw_prediction_results, hash_to_string_dict,
-                self.model.vocabs.target_vocab.special_words, topk=SHOW_TOP_CONTEXTS)
-            for raw_prediction, method_prediction in zip(raw_prediction_results, method_prediction_results):
-                print('Original name:\t' + method_prediction.original_name)
-                for name_prob_pair in method_prediction.predictions:
-                    print('\t(%f) predicted: %s' % (name_prob_pair['probability'], name_prob_pair['name']))
-                print('Attention:')
-                for attention_obj in method_prediction.attention_paths:
-                    print('%f\tcontext: %s,%s,%s' % (
-                    attention_obj['score'], attention_obj['token1'], attention_obj['path'], attention_obj['token2']))
-                if self.config.EXPORT_CODE_VECTORS:
-                    print('Code vector:')
-                    print(' '.join(map(str, raw_prediction.code_vector)))
+    def predict(self, path_folder):
+        files_list = list(Path(path_folder).glob('*.java'))
+        for input_filename in files_list:
+            file_name_str = str(input_filename)
+            # input_filename = Path('temp.java')
+            with open(file_name_str) as f:
+                code = f.read()
+                start = time.time()
+                predict_lines, hash_to_string_dict = self.path_extractor.extract_paths(code)
+                #with open('inspect.txt', 'w') as w:
+                    #w.write(inspect.getsource(self.model))
+                raw_prediction_results = self.model.predict(predict_lines)
+                end = time.time()
+                p = end - start
+                self.results[file_name_str] = p
+                print(f'Time passed: {p} for {file_name_str}')
+
+                method_prediction_results = common.parse_prediction_results(
+                    raw_prediction_results, hash_to_string_dict,
+                    self.model.vocabs.target_vocab.special_words, topk=SHOW_TOP_CONTEXTS)
+                for raw_prediction, method_prediction in zip(raw_prediction_results, method_prediction_results):
+                    print('Original name:\t' + method_prediction.original_name)
+                    for name_prob_pair in method_prediction.predictions:
+                        print('\t(%f) predicted: %s' % (name_prob_pair['probability'], name_prob_pair['name']))
+        total_time_for_one_iteration = sum(self.results.values())
+        times_arr = list(self.results.values())
+        print(f'Total time of running {len(files_list)} java methods is '
+              f'{total_time_for_one_iteration} secs for 1 iteration. \n'
+              f'Average time for 1 method: {mean(times_arr):0.3f} secs. \n'
+              f'Min time of 1 method: {min(times_arr):0.3f} secs, \n'
+              f'max time of 1 method: {max(times_arr):0.3f} secs, \n'
+              f'median: {median(times_arr):0.3f} secs, \n'
+              f'quantile 75%: {quantile(times_arr, 0.75):0.3f} secs, \n'
+              f'quantile 95%: {quantile(times_arr, 0.95):0.3f} secs')
+        all_res = json.dumps(self.results)
+        print(f'All results: {all_res}')