1) Do not load model every time when predicting
2) Create benchmark dataset
Results: 

Given 12 files with the following ncss: `[26,24,22,26,27,24,21,23, 25, 21,25]` (avg=24)

Total time of running 12 java methods is 5.19 secs.

Average time for 1 method: 0.433 secs. 
Min time of 1 method: 0.328 secs, 
max time of 1 method: 1.280 secs, 
median: 0.353 secs, 
quantile 75%: 0.370 secs, 
quantile 95%: 0.794 secs

| dataset/OracleSchema_prepareLookupStatement_23.0_8.0_7.0_24.java | dataset/ParseProjectXml_getFriends_21.0_10.0_10.0_26.java | dataset/ByteBufCodedInputStream_skipField_22.0_8.0_1.0_23.java | dataset/AbstractCSharpCodegen_toExampleValue_21.0_11.0_21.0_25.java | dataset/bcverify_createBitMaps_21.0_5.0_7.0_25.java | dataset/FastFourierTransform_multiply_22.0_8.0_10.0_24.java | dataset/EulerTotientFunction_primeFactorization_21.0_7.0_8.0_21.java | dataset/LitemallTopic_toString_21.0_1.0_1.0_22.java | dataset/temp.java  | dataset/AddContactDialog_actionPerformed_24.0_9.0_18.0_21.java | dataset/K6ClientCodegen_getAccept_21.0_7.0_15.0_26.java | dataset/FireBirdPlanParser_basicItem_24.0_5.0_3.0_27.java |
| ---------------------------------------------------------------- | --------------------------------------------------------- | -------------------------------------------------------------- | ------------------------------------------------------------------- | --------------------------------------------------- | ----------------------------------------------------------- | -------------------------------------------------------------------- | --------------------------------------------------- | ------------------ | -------------------------------------------------------------- | ------------------------------------------------------- | --------------------------------------------------------- |
| 1.2799160480499268                                               | 0.36394619941711426                                       | 0.34212398529052734                                            | 0.35977840423583984                                                 | 0.34549951553344727                                 | 0.39647889137268066                                         | 0.3414499759674072                                                   | 0.3685612678527832                                  | 0.3455014228820801 | 0.3282046318054199                                             | 0.3461117744445801                                      | 0.37454962730407715                                       |