@lyriccoder 2021-09-27
Here is the evaluation of prediction:
```
Time passed: 1.2799160480499268 for dataset/OracleSchema_prepareLookupStatement_23.0_8.0_7.0_24.java
Original name:	prepare|lookup|statement
	(0.337361) predicted: ['find', 'object']
	(0.193656) predicted: ['get', 'table']
	(0.105856) predicted: ['prepare', 'children', 'statement']
	(0.098230) predicted: ['prepare', 'objects', 'statement']
	(0.052807) predicted: ['create', 'function']
	(0.052362) predicted: ['create']
	(0.048819) predicted: ['prepare', 'lookup', 'statement']
	(0.040029) predicted: ['delete', 'property']
	(0.036268) predicted: ['fetch', 'child']
	(0.034612) predicted: ['create', 'statement']

Time passed: 0.36394619941711426 for dataset/ParseProjectXml_getFriends_21.0_10.0_10.0_26.java
Original name:	get|friends
	(0.243190) predicted: ['get', 'friends']
	(0.224343) predicted: ['execute']
	(0.101167) predicted: ['get', 'headers']
	(0.078209) predicted: ['get', 'targets']
	(0.068954) predicted: ['execute', 'batch']
	(0.063641) predicted: ['get', 'values']
	(0.062814) predicted: ['get', 'extensions']
	(0.056550) predicted: ['get', 'config']
	(0.050601) predicted: ['get', 'string', 'array']
	(0.050532) predicted: ['do', 'in', 'background']

Time passed: 0.34212398529052734 for dataset/ByteBufCodedInputStream_skipField_22.0_8.0_1.0_23.java
Original name:	skip|field
	(0.690057) predicted: ['skip', 'field']
	(0.083033) predicted: ['write', 'tag']
	(0.047170) predicted: ['write', 'fixed']
	(0.044408) predicted: ['write', 's', 'fixed']
	(0.029059) predicted: ['write', 'float']
	(0.027696) predicted: ['write', 'bool']
	(0.027053) predicted: ['write', 'int']
	(0.018330) predicted: ['read', 'bool']
	(0.017842) predicted: ['has', 'tag']
	(0.015352) predicted: ['merge', 'field', 'from']

Time passed: 0.35977840423583984 for dataset/AbstractCSharpCodegen_toExampleValue_21.0_11.0_21.0_25.java
Original name:	to|example|value
	(0.500784) predicted: ['map']
	(0.109229) predicted: ['parse', 'string']
	(0.075912) predicted: ['to', 'delimited', 'string']
	(0.062747) predicted: ['transform']
	(0.053820) predicted: ['change', 'case']
	(0.053187) predicted: ['translate']
	(0.041573) predicted: ['describe']
	(0.038286) predicted: ['get', 'plan']
	(0.032756) predicted: ['validate']
	(0.031707) predicted: ['apply']

Time passed: 0.34549951553344727 for dataset/bcverify_createBitMaps_21.0_5.0_7.0_25.java
Original name:	create|bit|maps
	(0.159308) predicted: ['check', 'identifier']
	(0.148416) predicted: ['start', 'children']
	(0.126237) predicted: ['clear']
	(0.125997) predicted: ['pre', 'process']
	(0.094271) predicted: ['find', 'field']
	(0.093978) predicted: ['case']
	(0.092243) predicted: ['reset', 'filter']
	(0.055023) predicted: ['randomize']
	(0.053283) predicted: ['reset']
	(0.051243) predicted: ['seen']

Time passed: 0.39647889137268066 for dataset/FastFourierTransform_multiply_22.0_8.0_10.0_24.java
Original name:	multiply
	(0.273076) predicted: ['map']
	(0.127115) predicted: ['cross']
	(0.104848) predicted: ['get', 'counters']
	(0.102239) predicted: ['join']
	(0.093057) predicted: ['div']
	(0.080413) predicted: ['encrypt']
	(0.057279) predicted: ['multiply']
	(0.054816) predicted: ['divide']
	(0.054183) predicted: ['encode', 'long']
	(0.052976) predicted: ['create']

Time passed: 0.3414499759674072 for dataset/EulerTotientFunction_primeFactorization_21.0_7.0_8.0_21.java
Original name:	prime|factorization
	(0.998947) predicted: ['get', 'factors']
	(0.000316) predicted: ['div']
	(0.000270) predicted: ['round', 'up']
	(0.000123) predicted: ['spawn']
	(0.000076) predicted: ['changed']
	(0.000067) predicted: ['mod']
	(0.000056) predicted: ['alloc']
	(0.000052) predicted: ['pow']
	(0.000049) predicted: ['calculate', 'size']
	(0.000045) predicted: ['atomic']

Time passed: 0.3685612678527832 for dataset/LitemallTopic_toString_21.0_1.0_1.0_22.java
Original name:	to|string
	(0.999520) predicted: ['to', 'string']
	(0.000159) predicted: ['get', 'description']
	(0.000118) predicted: ['get', 'name']
	(0.000046) predicted: ['to', 'debug', 'string']
	(0.000043) predicted: ['get', 'text']
	(0.000037) predicted: ['get', 'string']
	(0.000033) predicted: ['get', 'display', 'name']
	(0.000025) predicted: ['get', 'message']
	(0.000009) predicted: ['get', 'key']
	(0.000009) predicted: ['to', 'json']

Time passed: 0.3455014228820801 for dataset/temp.java
Original name:	create|new|gib|hub|issue
	(0.357877) predicted: ['create', 'issue']
	(0.252574) predicted: ['add', 'info']
	(0.095175) predicted: ['do', 'create']
	(0.063501) predicted: ['after', 'execute']
	(0.058984) predicted: ['test', 'create', 'issue']
	(0.041605) predicted: ['add', 'modifier']
	(0.037376) predicted: ['link']
	(0.033936) predicted: ['add', 'factory']
	(0.030194) predicted: ['start', 'wizard']
	(0.028778) predicted: ['create', 'message']

Time passed: 0.3282046318054199 for dataset/AddContactDialog_actionPerformed_24.0_9.0_18.0_21.java
Original name:	action|performed
	(0.999999) predicted: ['action', 'performed']
	(0.000001) predicted: ['do', 'action']
	(0.000000) predicted: ['perform', 'action']
	(0.000000) predicted: ['execute']
	(0.000000) predicted: ['handle']
	(0.000000) predicted: ['do', 'perform', 'action']
	(0.000000) predicted: ['mouse', 'clicked']
	(0.000000) predicted: ['fire', 'action', 'performed']
	(0.000000) predicted: ['button', 'clicked']
	(0.000000) predicted: ['handle', 'error']

Time passed: 0.3461117744445801 for dataset/K6ClientCodegen_getAccept_21.0_7.0_15.0_26.java
Original name:	get|accept
	(0.172223) predicted: ['resolve']
	(0.162860) predicted: ['authorize']
	(0.153323) predicted: ['add']
	(0.122905) predicted: ['write']
	(0.111956) predicted: ['get', 'content', 'type']
	(0.071126) predicted: ['accepts']
	(0.069017) predicted: ['handle']
	(0.055914) predicted: ['set', 'status']
	(0.042738) predicted: ['handle', 'get']
	(0.037937) predicted: ['put']

Time passed: 0.37454962730407715 for dataset/FireBirdPlanParser_basicItem_24.0_5.0_3.0_27.java
Original name:	basic|item
	(0.761872) predicted: ['execute']
	(0.081679) predicted: ['handle', 'message']
	(0.064570) predicted: ['visit']
	(0.022055) predicted: ['handle', 'request']
	(0.019528) predicted: ['accept']
	(0.016044) predicted: ['on', 'update']
	(0.013392) predicted: ['on', 'complete']
	(0.007029) predicted: ['on', 'receive']
	(0.006936) predicted: ['compare']
	(0.006896) predicted: ['on', 'bind', 'view', 'holder']
```