@lyriccoder 2021-09-27
[res.csv](https://github.com/tech-srl/code2vec/files/7237160/res.csv)
Here are results with java time calculation